forcing invite controller to render unregistered layout
this page will be viewed by both signed in as well as unsigned in members, but the layout follows style rules for use with only the unregistered layout so I'm forcing this layout and overriding the "choose_layout" method. alternatively, we could consider modifying the "choose_layout" method to detect if the controller is "invite" and force the unregistered layout therein.

Signed-off-by: Ben Liu - iMac <blzabub8@gmail.com>