update this view to use updated unregistered user layout properly
add cancel link to password reset page

add stubbed form for members to change their own passwords

I've embedded a form stub in the member profile view page so that a member can change their own password. We might want to consider simply making a link here to the devise/password/edit controller, or pull the devise password edit form into a partial and include the partial here. We must also answer whether we want password changes to be confirmed or not. I did not add a password confirmation field.

Signed-off-by: Ben Liu - iMac <blzabub8@gmail.com>